upgrade the simple WPF project add more CRUD methids in DL\BL for Student\Course\Lecturers

+++ DLAPI/DO/StudentInCourse.cs
@@ -10,6 +10,6 @@ namespace DO
     {
         public int PersonId { get; set; }
         public int CourseId { get; set; }
-        public int? Grade { get; set; }
+        public float? Grade { get; set; }
     }
 }
